Update and rename resumen de Lesser MP to resumen_Lesser_97.md
Just for kicks, I've made a couple of minuscule edits to the file name and format, as a suggestion. I think you can probably merge these into your branch before the pull request is approved - or choose not to merge them.

The change to the name is potentially important for organization. In general, it might be good to get used to saving files with names that have no spaces, and that have a defined structure to them (for these, i think it makes sense to have the author-year structure? might change in the future). Also, when you're working with raw files, you should define the 'file extension', which tells future users (and applications) what format you've written your text. In this case I added '.md' to suggest the 'markdown' format, which allowed me to add basic text formatting. I Italicized the journal name in your citation simply by adding asterisks around it, and thanks to the '.md' extension, GitHub knows automatically to interpret that. You could also just add '.txt' if you don't want to bother with the formatting - GitHub and other applications would then interpret my added asterisks as literal asterisks.