Create AnaColon_summary_Bingham_and_Ratcliff_2024
I wrote a summary of the complex eukaryotic evolution paper from a few weeks ago. I wanted to get comments on my summary in English to ensure I understood the paper before I try and write a summary in Spanish. There is a lot of complex vocabulary that I know in English but not in Spanish so that will also be a challenge.